added method MakeProduct

+++ database/migrations/2023_07_27_211510_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreign('caterogy_id')->references('id')->on('caterogy');
             $table->string('name');
             $table->float('price');
-            $table->integer('count')->default(0);
+            $table->float('count')->default(0);
             $table->unsignedBigInteger('type_id');
             $table->foreign('type_id')->references('id')->on('settings');
             $table->boolean('cyrrency')->default(0); // 0 - UZS 1 - USD
